The readFully function in CBOR_parsing now returns a size_t instead of an unsigned int.

+++ src/CBOR_parsing.cpp
@@ -229,7 +229,7 @@ bool expectTag(Reader r, uint64_t *v) {
 //  Read functions.
 // ***************************************************************************
 
-unsigned int readFully(Reader r, uint8_t *b, size_t len) {
+size_t readFully(Reader r, uint8_t *b, size_t len) {
   size_t count = 0;
   while (len > 0) {
     int read = r.readBytes(b, len);
